refactor(navigation): remove ThemeToggle component and update import
- Deleted the ThemeToggle component from components/home/ui
- Updated import path of ThemeToggle in Navigation component accordingly
- Simplified project structure by removing unused ThemeToggle component
